# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "The Audit Brief API",
    "version": "0.2.1",
...
    "contact": {
      "name": "Audit Brief Platform Team"
    },
    "license": {
      "name": "Proprietary — internal use only"
    }
  },
  "servers": [
    {
      "url": "https://uat.uno.wcgt.in/auditbrief",
      "description": "UAT / production-equivalent (basePath `/auditbrief`)."
    }
  ],
  "security": [
    {
      "cookieAuth": []
    }
  ],
...
  "paths": {
...
    "/api/search": {
      "get": {
        "tags": ["Search"],
        "operationId": "keywordSearch",
        "summary": "Keyword search over audit briefs",
        "description": "Public. Case-insensitive `contains` match on title / description / tags\nacross non-archived audit briefs. Response is **not** wrapped in `data`.\n",
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "example": "revenue recognition"
          },
          {
            "$ref": "#/components/parameters/PageParam"
          },
          {
            "$ref": "#/components/parameters/LimitParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Search hits.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KeywordSearchResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "post": {
        "tags": ["Search"],
        "operationId": "semanticSearch",
        "summary": "Semantic (pgvector) search over transcript text",
        "description": "Public. Generates an Azure OpenAI embedding for `query`, runs a cosine\nsimilarity search over `transcripts.embedding`, and returns the top 10\nresults with similarity > 0.7. Response is **not** wrapped in `data`.\n",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["query"],
                "properties": {
                  "query": {
                    "type": "string",
                    "minLength": 1
                  }
                }
              },
              "example": {
                "query": "how do we handle going-concern uncertainty?"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Semantic search hits.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SemanticSearchResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token",
        "description": "NextAuth v4 JWT session cookie. `httpOnly`, `SameSite=Lax`, and `Secure`\nin production. Over HTTPS the cookie is named\n`__Secure-next-auth.session-token`.\n\nIssued by `POST /api/auth/callback/credentials` (dev) or the Azure AD\nOAuth flow (`GET /api/auth/signin/azure-ad` → provider → `GET /api/auth/callback/azure-ad`).\nContains a signed JWT with `userId`, `role`, `email`, and `jti` (for\nserver-side revocation).\n"
      }
    },
...
    "parameters": {
...
      "PageParam": {
        "in": "query",
        "name": "page",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        },
        "description": "1-based page index."
      },
      "LimitParam": {
        "in": "query",
        "name": "limit",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        },
        "description": "Page size (max 100)."
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Malformed request, missing required fields, or invalid values.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "status": 400,
              "error_code": "BAD_REQUEST",
              "message": "key parameter required",
              "request_id": "3f2e7a6d-4b1a-4f9d-8a77-1b4c8e6d2a10"
            }
          }
        }
      },
...
      "InternalError": {
        "description": "Unhandled server error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "status": 500,
              "error_code": "INTERNAL_ERROR",
              "message": "Internal server error"
            }
          }
        }
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": ["status", "error_code", "message"],
        "properties": {
          "status": {
            "type": "integer",
            "description": "HTTP status code.",
            "example": 422
          },
          "error_code": {
            "type": "string",
            "enum": [
              "BAD_REQUEST",
              "UNAUTHORIZED",
              "FORBIDDEN",
              "NOT_FOUND",
              "CONFLICT",
              "VALIDATION_FAILED",
              "RATE_LIMITED",
              "INTERNAL_ERROR"
            ],
            "description": "Machine-readable error classification."
          },
          "message": {
            "type": "string",
            "description": "Human-readable message. Safe to surface to end-users."
          },
          "details": {
            "description": "Free-form diagnostic detail. For `VALIDATION_FAILED` this is a Zod\n`issues` array; for other codes it is commonly absent. May be null.\n"
          },
          "request_id": {
            "type": "string",
            "format": "uuid",
            "description": "Echo of the `x-request-id` header — use for log correlation."
          }
        }
      },
...
      "Domain": {
        "type": "string",
        "description": "Deduplicated union of podcast and learning-series knowledge domains.",
        "enum": [
          "Audit Methodology",
          "Accounting and Reporting",
          "Audit Technology",
          "Quality and Risk",
          "LEAP",
          "Auditing"
        ]
      },
...
      "KeywordSearchResponse": {
        "type": "object",
        "required": ["results", "query", "total", "page", "limit"],
        "properties": {
          "query": {
            "type": "string"
          },
          "total": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "results": {
            "type": "array",
            "items": {
              "type": "object",
              "required": ["id", "title", "domain"],
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "title": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "domain": {
                  "$ref": "#/components/schemas/Domain"
                },
                "tags": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "thumbnailUrl": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "SemanticSearchResponse": {
        "type": "object",
        "required": ["results", "query"],
        "properties": {
          "query": {
            "type": "string"
          },
          "results": {
            "type": "array",
            "description": "Top 10 transcript hits with similarity > 0.7, ordered by similarity desc.",
            "items": {
              "type": "object",
              "required": ["id", "auditBriefId", "auditBriefTitle", "content", "similarity"],
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid",
                  "description": "Transcript ID."
                },
                "auditBriefId": {
                  "type": "string",
                  "format": "uuid"
                },
                "auditBriefTitle": {
                  "type": "string"
                },
                "content": {
                  "type": "string",
                  "description": "Currently the full transcript text."
                },
                "startTime": {
                  "type": "number",
                  "description": "Reserved for future segment-level scoring. Always 0."
                },
                "endTime": {
                  "type": "number",
                  "description": "Reserved for future segment-level scoring. Always 0."
                },
                "similarity": {
                  "type": "number",
                  "description": "Cosine similarity in [0, 1]."
                }
              }
            }
          }
        }
      },
...
    }
  }
}